Check if person is logged in, improve removeOneContact method

+++ src/main/java/ui/controller/Delete.java
@@ -9,10 +9,13 @@ import javax.servlet.http.HttpSession;
 public class Delete extends RequestHandler {
     @Override
     public String handleRequest(HttpServletRequest request, HttpServletResponse response) {
+        HttpSession session = request.getSession();
+        Person person = (Person) session.getAttribute("person");
+
+        if (person == null) throw new RuntimeException("You need to be logged in to delete your account");
         String confirmation = request.getParameter("confirmation");
+
         if (!confirmation.isEmpty() && confirmation.equals("Delete")) {
-            HttpSession session = request.getSession();
-            Person person = (Person) session.getAttribute("person");
             getService().deletePerson(person);
             getService().removeContacts(person.getUserid());
             request.setAttribute("deleteMessage", "Your account has been removed!");
